test(vector_stores): split into one file per provider for maintainability
